adressed comments from hugo, naming and simpler code + adjusted test to refactoring from prev commit
